+++ src/components/SubscriptionSettingsMonero.vue
@@ -33,11 +33,36 @@
         <input type="number" id="price" v-model="subscriptionPrice" min="0.00" step="0.01">
         <span>XMR {{ $t('subscriptions.price_per_month') }}</span>
       </div>
+      <div
+        v-if="availablePaymentMethods.length > 1"
+        class="method-input-group"
+      >
+        <label for="payment_method">{{ $t('subscriptions.payment_type') }}</label>
+        <select
+          id="payment_method"
+          v-model="paymentMethod"
+        >
+          <option
+            v-for="method in availablePaymentMethods"
+            :key="method.name"
+            :value="method"
+          >
+            {{ method.name }}
+          </option>
+        </select>
+      </div>
       <input
         type="text"
         id="payout_address"
         v-model="subscriptionPayoutAddress"
         :placeholder="$t('subscriptions.payout_address')"
+      >
+      <input
+        v-if="paymentMethod?.type === 'monero-light'"
+        type="text"
+        id="view_key"
+        v-model="viewKey"
+        placeholder="View key"
       >
       <button
         type="submit"
@@ -60,6 +85,7 @@
 
 <script setup lang="ts">
 import { computed, onMounted, ref } from "vue"
+import { useI18n } from "vue-i18n"
 import { useRouter } from "vue-router"
 
 import {
@@ -77,6 +103,13 @@ import { useActorHandle } from "@/composables/handle"
 import { useInstanceInfo } from "@/composables/instance"
 import { useCurrentUser } from "@/composables/user"
 
+interface PaymentMethod {
+  type: "monero" | "monero-light",
+  name: string,
+  chainId: string,
+}
+
+const { t } = useI18n({ useScope: "global" })
 const router = useRouter()
 const { getActorLocation } = useActorHandle()
 const {
@@ -84,19 +117,42 @@ const {
   ensureCurrentUser,
   setCurrentUser,
 } = useCurrentUser()
-const { getBlockchainInfo } = useInstanceInfo()
+const { instance, getBlockchainInfo } = useInstanceInfo()
 
 const isLoading = ref(false)
 const subscriptionOption = ref<SubscriptionOption | null>(null)
 const subscriptionOptionLoaded = ref(false)
 
+const paymentMethod = ref<PaymentMethod | undefined>(undefined)
 const subscriptionPrice = ref(0.01)
 const subscriptionPayoutAddress = ref("")
+const viewKey = ref("")
 const isFormVisible = ref(false)
 const errorMessage = ref<string | null>(null)
 
 const subscriberCount = computed(() => {
   return ensureCurrentUser().subscribers_count
+})
+
+const availablePaymentMethods = computed<PaymentMethod[]>(() => {
+  if (!instance.value) {
+    return []
+  }
+  const blockchain = getBlockchainInfo()
+  if (!blockchain) {
+    return []
+  }
+  const mainChainId = blockchain.chain_id
+  return instance.value.blockchains
+    // Only show payment methods with the same chain ID
+    .filter(config => config.chain_id === mainChainId)
+    .map(config => {
+      return {
+        type: config.chain_metadata?.is_forwarding_required ? "monero" : "monero-light",
+        name: config.chain_metadata?.is_forwarding_required ? t("subscriptions.monero_forwarded") : t("subscriptions.monero_direct_requires_a_view_key"),
+        chainId: config.chain_id,
+      }
+    })
 })
 
 onMounted(async () => {
@@ -111,8 +167,14 @@ onMounted(async () => {
 async function loadSubscriptionSettings() {
   const subscriptionOptions = await getSubscriptionOptions(ensureAuthToken())
   subscriptionOption.value = subscriptionOptions.find((item) => {
-    return item.type === "monero"
+    return item.type === "monero" || item.type === "monero-light"
   }) || null
+  if (subscriptionOption.value === null) {
+    paymentMethod.value = availablePaymentMethods.value[0]
+  } else {
+    paymentMethod.value = availablePaymentMethods.value
+      .find(method => method.type === subscriptionOption.value?.type)
+  }
   subscriptionOptionLoaded.value = true
   if (
     subscriptionOption.value?.price &&
@@ -121,7 +183,10 @@ async function loadSubscriptionSettings() {
     subscriptionPrice.value = getPricePerMonth(subscriptionOption.value.price)
     subscriptionPayoutAddress.value = subscriptionOption.value.payout_address
   }
-  if (subscriptionOption.value === null) {
+  if (subscriptionOption.value?.view_key) {
+     viewKey.value = subscriptionOption.value.view_key
+  }
+  if (subscriptionOption.value === null && paymentMethod.value !== undefined) {
     isFormVisible.value = true
   }
 }
@@ -145,8 +210,7 @@ function isFormValid(): boolean {
 }
 
 async function saveSubscriptionSettings() {
-  const blockchain = getBlockchainInfo()
-  if (blockchain === null) {
+  if (paymentMethod.value === undefined) {
     return
   }
   isLoading.value = true
@@ -154,9 +218,11 @@ async function saveSubscriptionSettings() {
   try {
     user = await registerMoneroSubscriptionOption(
       ensureAuthToken(),
-      blockchain.chain_id,
+      paymentMethod.value.type,
+      paymentMethod.value.chainId,
       getPricePerSec(subscriptionPrice.value),
       subscriptionPayoutAddress.value,
+      viewKey.value,
     )
   } catch (error: any) {
     isLoading.value = false
@@ -225,6 +291,18 @@ form {
 
   input[type="number"] {
     width: 100px;
+  }
+}
+
+.method-input-group {
+  align-items: center;
+  display: flex;
+  flex-wrap: wrap;
+  gap: $input-padding;
+  justify-content: center;
+
+  label {
+    font-weight: bold;
   }
 }