+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Don't change toggled content warning button on hover.
+
 ## [4.14.0] - 2025-11-28
 
 ### Added