+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Made follower removal confirmation translatable.
 - Display custom emojis in greentext.
 
+### Removed
+
+- Removed support for token gate.
+
 ### Fixed
 
 - Don't change toggled content warning button on hover.