+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Use Instance API v2.
 - Change "Publish" to "Send" if post visibility is "Direct".
+- Updated translations ([#147](https://codeberg.org/silverpill/mitra-web/pulls/147)).
 
 ## [4.15.0] - 2025-12-22